Update Pool for better usability
