+++ StubbUnity/StubbUnity/Src/StubbFramework/Pooling/Pools.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Runtime.CompilerServices;
 
 namespace StubbUnity.StubbFramework.Pooling
@@ -8,76 +7,71 @@ namespace StubbUnity.StubbFramework.Pooling
     public class Pools
     {
         #region Singleton
-        private static Pools _instance;  
+
+        private static Pools _instance;
         public static Pools I => _instance ?? (_instance = new Pools());
+
         #endregion
 
         private readonly Dictionary<Type, IPoolGeneric> _pools;
 
         public int NumPools => _pools.Count;
-        
+
         private Pools()
         {
             _pools = new Dictionary<Type, IPoolGeneric>();
         }
 
         /// <summary>
-        /// Creates a pool with given type.
+        /// Get or creates and get a pool with given type.
         /// </summary>
-        /// <param name="capacity">Initial capacity of the pool. Min value is 10.</param>
-        public IPool<T> Create<T>(int capacity)
+        /// <param name="capacity">Initial capacity of the pool.</param>
+        /// <param name="prewarm">If 'true' will be created the number of instances equal to the 'capacity' of the pool.</param>
+        public IPool<T> Get<T>(int capacity = 5, bool prewarm = false)
         {
-            _CheckIfPoolAlreadyExist<T>();
+            var pool = _Create<T>(capacity);
 
-            var pool = new Pool<T>(capacity);
-            pool.OnRemove += OnRemovePoolHandler;
-            _pools[typeof(T)] = pool;
-            
+            if (pool.Creator == null && pool.CreateMethod == null)
+                pool.CreateMethod = () => (T) Activator.CreateInstance(typeof(T), true);
+
+            if (prewarm)
+                pool.PreWarm(pool.Size);
+
             return pool;
         }
-        
+
         /// <summary>
-        /// Creates a pool with given type.
+        /// Get or creates and get a pool with given type.
         /// </summary>
-        /// <param name="capacity">Initial capacity of the pool. Min value is 10.</param>
+        /// <param name="capacity">Initial capacity of the pool.</param>
         /// <param name="creator">Instance of ICreator which will be used for creating an instance of pool's type.</param>
         /// <param name="prewarm">If 'true' will be created the number of instances equal to the 'capacity' of the pool.</param>
-        public IPool<T> Create<T>(int capacity, ICreator<T> creator, bool prewarm = false)
+        public IPool<T> Get<T>(int capacity, ICreator<T> creator, bool prewarm = false)
         {
-            var pool = Create<T>(capacity);
+            var pool = _Create<T>(capacity);
             pool.Creator = creator;
-            
-            if (prewarm) 
+
+            if (prewarm)
                 pool.PreWarm(pool.Size);
 
             return pool;
         }
 
         /// <summary>
-        /// Creates a pool with given type.
+        /// Get or creates and get a pool with given type.
         /// </summary>
-        /// <param name="capacity">Initial capacity of the pool. Min value is 10.</param>
+        /// <param name="capacity">Initial capacity of the pool.</param>
         /// <param name="createMethod">Method which will be used for creating an instance of pool's type.</param>
         /// <param name="prewarm">If 'true' will be created the number of instances equal to the 'capacity' of the pool.</param>
-        public IPool<T> Create<T>(int capacity, Func<T> createMethod, bool prewarm = false)
+        public IPool<T> Get<T>(int capacity, Func<T> createMethod, bool prewarm = false)
         {
-            var pool = Create<T>(capacity);
+            var pool = _Create<T>(capacity);
             pool.CreateMethod = createMethod;
 
-            if (prewarm) 
+            if (prewarm)
                 pool.PreWarm(pool.Size);
 
             return pool;
-        }
-
-        /// <summary>
-        /// Returns a pool by given type.
-        /// </summary>
-        public IPool<T> Get<T>()
-        {
-            _CheckIfPoolDoesntExist<T>();
-
-            return _pools[typeof(T)] as Pool<T>;
         }
 
         /// <summary>
@@ -102,7 +96,7 @@ namespace StubbUnity.StubbFramework.Pooling
                 pair.Value.OnRemove -= OnRemovePoolHandler;
                 pair.Value.Dispose();
             }
-            
+
             _pools.Clear();
         }
 
@@ -121,23 +115,21 @@ namespace StubbUnity.StubbFramework.Pooling
         {
             _pools.Remove(type);
         }
-
-        [Conditional("DEBUG")]
-        private void _CheckIfPoolDoesntExist<T>()
-        {
-            if (!Has<T>())
-            {
-                throw new Exception($"Pool with type {typeof(T)} doesn't exist! Make sure you register before use.");
-            }
-        }
-
-        [Conditional("DEBUG")]
-        private void _CheckIfPoolAlreadyExist<T>()
+        
+        /// <summary>
+        /// Creates a pool with given type.
+        /// </summary>
+        /// <param name="capacity">Initial capacity of the pool. Min value is 5.</param>
+        private IPool<T> _Create<T>(int capacity = 5)
         {
             if (Has<T>())
-            {
-                throw new Exception($"Pool with type {typeof(T)} already exist!");
-            }
+                return _pools[typeof(T)] as Pool<T>;
+
+            var pool = new Pool<T>(capacity);
+            pool.OnRemove += OnRemovePoolHandler;
+            _pools[typeof(T)] = pool;
+
+            return pool;
         }
     }
 }
